Update `AlphaCard` docs
