+++ .changeset/five-bats-attend.md
@@ -1,0 +1,5 @@
+---
+'polaris.shopify.com': patch
+---
+
+Update `AlphaCard` component docs
