--- polaris.shopify.com/pages/examples/alpha-card-subdued.tsx
@@ -1,18 +1,0 @@
-import {AlphaCard, Text, AlphaStack} from '@shopify/polaris';
-import React from 'react';
-import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
-
-function AlphaCardExample() {
-  return (
-    <AlphaCard background="surface-subdued">
-      <AlphaStack spacing="5">
-        <Text as="h3" variant="headingMd">
-          Online store dashboard
-        </Text>
-        <p>View a summary of your online store’s performance.</p>
-      </AlphaStack>
-    </AlphaCard>
-  );
-}
-
-export default withPolarisExample(AlphaCardExample);
